# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogoutResponse": {
        "properties": {
          "logout_url": {
            "description": "URL the frontend should navigate to for WorkOS session logout.",
            "title": "Logout Url",
            "type": "string"
          },
          "redirect_to": {
            "description": "Allowed frontend URL WorkOS should return to after logout.",
            "title": "Redirect To",
            "type": "string"
          }
        },
        "required": [
          "logout_url",
          "redirect_to"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
...
      "OrganizationSwitch": {
        "properties": {
          "organization_id": {
            "description": "Organization id to make active in the sealed WorkOS session.",
            "title": "Organization Id",
            "type": "string"
          }
        },
        "required": [
          "organization_id"
        ],
        "title": "OrganizationSwitch",
        "type": "object"
      },
      "OrganizationSwitchResponse": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "workos_organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workos Organization Id"
          }
        },
        "title": "OrganizationSwitchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFHeader": {
        "description": "Signed double-submit CSRF token required for unsafe browser requests in WorkOS auth mode.",
        "in": "header",
        "name": "x-archagent-csrf",
        "type": "apiKey"
      },
      "CloudTasksOIDC": {
        "description": "Cloud Run OIDC token presented by Google Cloud Tasks in production.",
        "scheme": "bearer",
        "type": "http"
      },
      "CollectorBearer": {
        "description": "Collector credential returned by /collector/v1/register or /collector/v1/credentials/rotate.",
        "scheme": "bearer",
        "type": "http"
      },
      "DevelopmentIdentityHeader": {
        "description": "Local development/test identity selector. Disabled in production.",
        "in": "header",
        "name": "x-archagent-user",
        "type": "apiKey"
      },
      "WorkOSSessionCookie": {
        "description": "Sealed WorkOS AuthKit session cookie used in hosted auth mode.",
        "in": "cookie",
        "name": "__Host-archagent-session",
        "type": "apiKey"
      },
      "WorkOSWebhookSignature": {
        "description": "WorkOS webhook signature header validated with the configured webhook secret.",
        "in": "header",
        "name": "workos-signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nArchAgent API contract for browser frontend, collector, staff, webhook, and worker integrations.\n\nUse `/openapi.json` as the machine-readable source of truth. Swagger UI is available at `/docs`\nand ReDoc is available at `/redoc` when the FastAPI app is running.\n",
    "title": "ArchAgent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/callback": {
      "get": {
        "description": "Exchange a WorkOS authorization code, mirror the identity locally, set the sealed session cookie, and redirect to the frontend.",
        "operationId": "callback_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "description": "Frontend return URL previously passed through the WorkOS OAuth state.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Frontend return URL previously passed through the WorkOS OAuth state.",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete WorkOS Login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/login": {
      "get": {
        "description": "Redirect the browser to WorkOS AuthKit hosted login.",
        "operationId": "login_auth_login_get",
        "parameters": [
          {
            "description": "Allowed frontend URL to return to after authentication.",
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Allowed frontend URL to return to after authentication.",
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start WorkOS Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Clear the ArchAgent session and CSRF cookies, then redirect through the WorkOS logout URL.",
        "operationId": "logout_auth_logout_post",
        "parameters": [
          {
            "description": "Allowed frontend URL to return to after logout.",
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Allowed frontend URL to return to after logout.",
              "title": "Return To"
            }
          },
          {
            "description": "Use json when logout is initiated by fetch and the frontend will navigate to logout_url.",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "redirect",
              "description": "Use json when logout is initiated by fetch and the frontend will navigate to logout_url.",
              "enum": [
                "redirect",
                "json"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Logout URL payload for fetch-based clients."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          }
        ],
        "summary": "Log Out",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/v1/session/organization": {
      "post": {
        "description": "Refresh the sealed WorkOS session for another organization available to the current user.",
        "operationId": "switch_organization_v1_session_organization_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationSwitch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSwitchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CSRFHeader": [],
            "WorkOSSessionCookie": []
          },
          {
            "DevelopmentIdentityHeader": []
          }
        ],
        "summary": "Switch Active Organization",
        "tags": [
          "Auth",
          "Product"
        ]
      }
    },
...
  },
...
}